backend/requirements.txt
alembic==1.16.5
firebase-admin==7.1.0
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
google-genai==1.47.0
gunicorn==23.0.0
pytest==8.4.2
redis==7.0.1
